binja: move the stack string detection to the function level. Fix https://github.com/mandiant/capa/issues/2516
